+++ flask_typed_routes/__init__.py
@@ -1,7 +1,6 @@
-from flask_typed_routes.app import FlaskTypedRoutes  # noqa
+from flask_typed_routes.app import FlaskTypedRoutes, Mode, typed_route  # noqa
 from flask_typed_routes.errors import ValidationError  # noqa
 from flask_typed_routes.fields import Query, Path, Header, Cookie, Body  # noqa
-from flask_typed_routes.core import typed_route  # noqa
 
 __all__ = [
     "ValidationError",
@@ -12,4 +11,5 @@ __all__ = [
     "Body",
     "typed_route",
     "FlaskTypedRoutes",
+    "Mode",
 ]