+++ tests/unit/test_fields.py
@@ -9,24 +9,24 @@ import flask_typed_routes.errors as ftr_errors
 import flask_typed_routes.fields as ftr_fields
 
 
-def test_path_field(flask_app):
-    with flask_app.test_request_context('/'):
+def test_path_field(flask_app_auto):
+    with flask_app_auto.test_request_context('/'):
         flask.request.view_args = {'id': 123}
         assert ftr_fields.Path(alias='id').value == 123
 
 
 @pytest.mark.parametrize('multi', [True, False])
-def test_query_field(flask_app, multi):
+def test_query_field(flask_app_auto, multi):
     expected = ['term'] if multi else 'term'
-    with flask_app.test_request_context('/?search=term'):
+    with flask_app_auto.test_request_context('/?search=term'):
         result = ftr_fields.Query(alias='search', multi=multi).value
         assert result == expected
 
 
 @pytest.mark.parametrize('multi', [True, False])
-def test_cookie_field(flask_app, multi):
+def test_cookie_field(flask_app_auto, multi):
     expected = ['abc123', 'abc124'] if multi else 'abc123'
-    with flask_app.test_request_context('/'):
+    with flask_app_auto.test_request_context('/'):
         headers = [('Cookie', 'session_id=abc123; session_id=abc124')]
         builder = werkzeug.test.EnvironBuilder(path='/', headers=headers)
         env = builder.get_environ()
@@ -37,25 +37,25 @@ def test_cookie_field(flask_app, multi):
         assert result == expected
 
 
-def test_header_field(flask_app):
+def test_header_field(flask_app_auto):
     expected = 'Bearer token1'
     headers = [
         ('Authorization', 'Bearer token1'),
     ]
     # Fixed: The `headers` multi dont work
-    with flask_app.test_request_context('/', headers=headers):
+    with flask_app_auto.test_request_context('/', headers=headers):
         result = ftr_fields.Header(alias='Authorization').value
         assert result == expected
 
 
-def test_body_field(flask_app):
-    with flask_app.test_request_context('/', json={'key': 'value'}):
+def test_body_field(flask_app_auto):
+    with flask_app_auto.test_request_context('/', json={'key': 'value'}):
         result = ftr_fields.Body().value
     assert result == {'key': 'value'}
 
 
-def test_body_embed_field(flask_app):
-    with flask_app.test_request_context('/', json={'key': {"subkey": "value"}}):
+def test_body_embed_field(flask_app_auto):
+    with flask_app_auto.test_request_context('/', json={'key': {"subkey": "value"}}):
         result = ftr_fields.Body(alias="key", embed=True).value
     assert result == {"subkey": "value"}